refactor: organize specs into folders
