+++ client/vitest.config.ts
@@ -16,7 +16,6 @@ export default mergeConfig(
         exclude: ['**/views/**', ''],
         reporter: ['text', 'html'],
       },
-      setupFiles: ['./tests/vitest-setup.ts'],
     },
   })
 );
